feat: health resource

+++ Assets/Scripts/Runtime/ResourceObject.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Resource")]
 public class ResourceObject : ScriptableObject
 {
-    [SerializeField] internal int consumptionPerDay;
+    [SerializeField] internal float consumptionPerDay;
 }
